+++ src/components/DataDisplay/Chip/stories/Chip.stories.tsx
@@ -4,8 +4,33 @@ import Chip from "./Chip";
 
 const Template: ComponentStory<typeof Chip> = (args) => (
   <Stack direction="row" spacing={2} alignItems="center" justifyContent="center" height="100%">
+    <Chip size="xSmall" label="xSmall" {...args} />
     <Chip size="small" label="small" {...args} />
     <Chip size="medium" label="medium" {...args} />
+  </Stack>
+);
+
+const TemplateColor: ComponentStory<typeof Chip> = (args) => (
+  <Stack direction="row" spacing={2} alignItems="center" justifyContent="center" height="100%">
+    <Stack spacing={1}>
+      <Chip variant="outlined" label="outlined default" color="default" {...args} />
+      <Chip variant="outlined" label="outlined primary" color="primary" {...args} />
+      <Chip variant="outlined" label="outlined secondary" color="secondary" {...args} />
+      <Chip variant="outlined" label="outlined info" color="info" {...args} />
+      <Chip variant="outlined" label="outlined success" color="success" {...args} />
+      <Chip variant="outlined" label="outlined warning" color="warning" {...args} />
+      <Chip variant="outlined" label="outlined error" color="error" {...args} />
+    </Stack>
+
+    <Stack spacing={1}>
+      <Chip label="default" color="default" {...args} />
+      <Chip label="primary" color="primary" {...args} />
+      <Chip label="secondary" color="secondary" {...args} />
+      <Chip label="info" color="info" {...args} />
+      <Chip label="success" color="success" {...args} />
+      <Chip label="warning" color="warning" {...args} />
+      <Chip label="error" color="error" {...args} />
+    </Stack>
   </Stack>
 );
 
@@ -30,6 +55,11 @@ Rounded.args = {
 export const WithDeleteIcon = Template.bind({});
 WithDeleteIcon.args = {
   onDelete: () => {},
+};
+
+export const Colors = TemplateColor.bind({});
+Colors.args = {
+  onDelete: undefined,
 };
 
 export default {