+++ src/constants/fonts.ts
@@ -1,0 +1,2 @@
+export const defaultFontWeight = [400, 500, 600, 700];
+export const defaultFontFamily = "Poppins, sans-serif";